feat: update LongMessageScheduleController to return structured API response and refactor request DTO
