+++ src/main/java/oneclass/oneclass/domain/sendon/sms/controller/LongMessageScheduleController.java
@@ -1,8 +1,10 @@
 package oneclass.oneclass.domain.sendon.sms.controller;
 
 import lombok.RequiredArgsConstructor;
-import oneclass.oneclass.domain.sendon.sms.dto.LongMessageScheduleRequest;
+import oneclass.oneclass.domain.sendon.sms.dto.request.LongMessageScheduleRequest;
+import oneclass.oneclass.domain.sendon.sms.dto.response.LongMessageScheduleResponse;
 import oneclass.oneclass.domain.sendon.sms.longmessage.SmsSendLongMessageSchedule;
+import oneclass.oneclass.global.dto.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,10 +19,10 @@ public class LongMessageScheduleController {
 
     // @param reservation 예약 시간 (yyyy-MM-dd HH:mm:ss)
     @PostMapping("/lms/schedule")
-    public ResponseEntity<?> sendLongMessage(
+    public ResponseEntity<ApiResponse<LongMessageScheduleResponse>> sendLongMessage(
             @RequestBody @Validated LongMessageScheduleRequest dto
     ) {
         smsSendLongMessageScheduleScenario.send(dto.getMessage(), dto.getTitle(), dto.getReservation().toString());
-        return ResponseEntity.ok("예약 LMS 발송 요청 완료 (" + dto.getReservation() + ")");
+        return ResponseEntity.ok(ApiResponse.success(new LongMessageScheduleResponse(dto.getMessage(), dto.getTitle(), dto.getReservation())));
     }
 }
