+++ src/main/java/oneclass/oneclass/domain/sendon/sms/dto/request/LongMessageScheduleRequest.java
@@ -1,4 +1,4 @@
-package oneclass.oneclass.domain.sendon.sms.dto;
+package oneclass.oneclass.domain.sendon.sms.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -8,7 +8,6 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDateTime;
 
-// DTO 정의
 @Setter
 @Getter
 public class LongMessageScheduleRequest {
