+++ src/main/java/oneclass/oneclass/domain/sendon/sms/dto/response/LongMessageScheduleResponse.java
@@ -1,0 +1,27 @@
+package oneclass.oneclass.domain.sendon.sms.dto.response;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDateTime;
+
+@Setter
+@Getter
+@AllArgsConstructor
+public class LongMessageScheduleResponse {
+
+    @NotBlank
+    private String message;
+
+    @NotBlank
+    private String title;
+
+    @NotNull
+    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime reservation;
+
+}
